+++ src/c++/perf_analyzer/infer_context.cc
@@ -24,7 +24,7 @@
 // (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 // OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-#include "infer_manager.h"
+#include "infer_context.h"
 
 namespace triton { namespace perfanalyzer {
